@@ -1,9 +1,10 @@
 package com.aastudio.sarollahi.moviebox.Activities;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.aastudio.sarollahi.moviebox.Data.DbHelper;
 import com.aastudio.sarollahi.moviebox.Data.SeenMoviesRecyclerViewAdapter;
@@ -37,7 +38,7 @@ public class SeenMoviesActivity extends AppCompatActivity {
 
 
     @Override
-    public boolean onSupportNavigateUp(){
+    public boolean onSupportNavigateUp() {
         finish();
         return true;
     }
@@ -53,7 +54,7 @@ public class SeenMoviesActivity extends AppCompatActivity {
     private void getmovies() {
 
         db = new DbHelper(this);
-        recyclerView = (RecyclerView) findViewById(R.id.recyclerViewID);
+        recyclerView = findViewById(R.id.recyclerViewID);
         recyclerView.setHasFixedSize(true);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
 
