@@ -1,9 +1,9 @@
 package com.aastudio.sarollahi.moviebox.Data;
 
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,6 +17,7 @@ public class ViewPagerAdapter extends FragmentStatePagerAdapter {
     public ViewPagerAdapter(FragmentManager fm) {
         super(fm);
     }
+
 
     @Override
     public Fragment getItem(int i) {
@@ -35,8 +36,7 @@ public class ViewPagerAdapter extends FragmentStatePagerAdapter {
     }
 
 
-
-    public void AddFragment (Fragment fragment, String title) {
+    public void AddFragment(Fragment fragment, String title) {
 
         lstFragment.add(fragment);
         lstTitles.add(title);
