@@ -1,14 +1,19 @@
 package com.aastudio.sarollahi.moviebox.Activities;
 
 import android.annotation.SuppressLint;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
 import android.view.View;
 import android.view.WindowManager;
+import android.widget.Button;
+import android.widget.LinearLayout;
 import android.widget.ProgressBar;
+import android.widget.RelativeLayout;
+import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.aastudio.sarollahi.moviebox.Data.PersonRecyclerViewAdapter;
 import com.aastudio.sarollahi.moviebox.Model.Actor;
@@ -21,6 +26,14 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.Volley;
+import com.facebook.ads.Ad;
+import com.facebook.ads.AdError;
+import com.facebook.ads.AdIconView;
+import com.facebook.ads.AdOptionsView;
+import com.facebook.ads.NativeAdLayout;
+import com.facebook.ads.NativeAdListener;
+import com.facebook.ads.NativeBannerAd;
+import com.google.android.material.snackbar.Snackbar;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -36,22 +49,23 @@ public class SearchByActorActivity extends AppCompatActivity {
     private List<Actor> actorList;
     private RequestQueue queue;
     private ProgressBar mProgressBar;
+    private NativeBannerAd nativeBannerAd;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_search_by_actor);
 
-        mProgressBar = (ProgressBar) findViewById(R.id.progressBaract);
+        mProgressBar = findViewById(R.id.progressBaract);
 
         assert getSupportActionBar() != null;   //null check
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
         queue = Volley.newRequestQueue(this);
 
-        int mNoOfColumns = Utility.calculateNoOfColumns(getApplicationContext(),85);
+        int mNoOfColumns = Utility.calculateNoOfColumns(getApplicationContext(), 85);
 
-        recyclerView = (RecyclerView) findViewById(R.id.recyclerView);
+        recyclerView = findViewById(R.id.recyclerView);
         recyclerView.setHasFixedSize(true);
         recyclerView.setLayoutManager(new GridLayoutManager(this, mNoOfColumns));
 
@@ -64,17 +78,137 @@ public class SearchByActorActivity extends AppCompatActivity {
         // getMovies(search);
         actorList = getActor(search);
 
-        personRecyclerViewAdapter = new PersonRecyclerViewAdapter(this, actorList );
+        personRecyclerViewAdapter = new PersonRecyclerViewAdapter(this, actorList);
         recyclerView.setAdapter(personRecyclerViewAdapter);
         personRecyclerViewAdapter.notifyDataSetChanged();
 
 
+        nativeBannerAd = new NativeBannerAd(this, Constants.Actor_ZONE_ID);
+        nativeBannerAd.setAdListener(new NativeAdListener() {
 
+            @Override
+            public void onMediaDownloaded(Ad ad) {
+
+            }
+
+            @Override
+            public void onError(Ad ad, AdError adError) {
+
+            }
+
+            @Override
+            public void onAdLoaded(Ad ad) {
+                // Race condition, load() called again before last ad was displayed
+                if (nativeBannerAd == null || nativeBannerAd != ad) {
+                    return;
+                }
+                // Inflate Native Banner Ad into Container
+                inflateAd(nativeBannerAd);
+                List<View> clickableViews = new ArrayList<>();
+            }
+
+            @Override
+            public void onAdClicked(Ad ad) {
+
+            }
+
+            @Override
+            public void onLoggingImpression(Ad ad) {
+
+            }
+
+        });
+        // load the ad
+        nativeBannerAd.loadAd();
 
     }
 
+    private void inflateAd(NativeBannerAd nativeBannerAd) {
+        // Unregister last ad
+        nativeBannerAd.unregisterView();
+
+        // Add the Ad view into the ad container.
+        NativeAdLayout nativeAdLayout = findViewById(R.id.native_banner_ad_container);
+        LayoutInflater inflater = LayoutInflater.from(SearchByActorActivity.this);
+        // Inflate the Ad view.  The layout referenced is the one you created in the last step.
+        LinearLayout adView = (LinearLayout) inflater.inflate(R.layout.native_banner_ad_layout, nativeAdLayout, false);
+        nativeAdLayout.addView(adView);
+
+        // Add the AdChoices icon
+        RelativeLayout adChoicesContainer = adView.findViewById(R.id.ad_choices_container);
+        AdOptionsView adOptionsView = new AdOptionsView(SearchByActorActivity.this, nativeBannerAd, nativeAdLayout);
+        adChoicesContainer.removeAllViews();
+        adChoicesContainer.addView(adOptionsView, 0);
+
+        // Create native UI using the ad metadata.
+        TextView nativeAdTitle = adView.findViewById(R.id.native_ad_title);
+        TextView nativeAdSocialContext = adView.findViewById(R.id.native_ad_social_context);
+        TextView sponsoredLabel = adView.findViewById(R.id.native_ad_sponsored_label);
+        AdIconView nativeAdIconView = adView.findViewById(R.id.native_icon_view);
+        Button nativeAdCallToAction = adView.findViewById(R.id.native_ad_call_to_action);
+
+        // Set the Text.
+        nativeAdCallToAction.setText(nativeBannerAd.getAdCallToAction());
+        nativeAdCallToAction.setVisibility(
+                nativeBannerAd.hasCallToAction() ? View.VISIBLE : View.INVISIBLE);
+        nativeAdTitle.setText(nativeBannerAd.getAdvertiserName());
+        nativeAdSocialContext.setText(nativeBannerAd.getAdSocialContext());
+        sponsoredLabel.setText(nativeBannerAd.getSponsoredTranslation());
+
+        // Register the Title and CTA button to listen for clicks.
+        List<View> clickableViews = new ArrayList<>();
+        clickableViews.add(nativeAdTitle);
+        clickableViews.add(nativeAdCallToAction);
+        nativeBannerAd.registerViewForInteraction(adView, nativeAdIconView, clickableViews);
+    }
+
     @Override
-    public boolean onSupportNavigateUp(){
+    protected void onResume() {
+        super.onResume();
+        try {
+            nativeBannerAd = new NativeBannerAd(this, Constants.Actor_ZONE_ID);
+            nativeBannerAd.setAdListener(new NativeAdListener() {
+
+                @Override
+                public void onMediaDownloaded(Ad ad) {
+
+                }
+
+                @Override
+                public void onError(Ad ad, AdError adError) {
+
+                }
+
+                @Override
+                public void onAdLoaded(Ad ad) {
+                    // Race condition, load() called again before last ad was displayed
+                    if (nativeBannerAd == null || nativeBannerAd != ad) {
+                        return;
+                    }
+                    // Inflate Native Banner Ad into Container
+                    inflateAd(nativeBannerAd);
+                    List<View> clickableViews = new ArrayList<>();
+                }
+
+                @Override
+                public void onAdClicked(Ad ad) {
+
+                }
+
+                @Override
+                public void onLoggingImpression(Ad ad) {
+
+                }
+
+            });
+            // load the ad
+            nativeBannerAd.loadAd();
+        } catch (Error error) {
+        }
+    }
+
+    @Override
+    public boolean onSupportNavigateUp() {
         finish();
         return true;
     }
@@ -88,12 +222,12 @@ public class SearchByActorActivity extends AppCompatActivity {
         mProgressBar.setVisibility(View.VISIBLE);
 
         JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET,
-                Constants.MAIN_FIND_ACROT_BY_NAME + name,null, new Response.Listener<JSONObject>() {
+                Constants.MAIN_FIND_ACROT_BY_NAME + name, null, new Response.Listener<JSONObject>() {
             @SuppressLint("ResourceType")
             @Override
             public void onResponse(JSONObject response) {
 
-                try{
+                try {
                     JSONArray actorsArray = response.getJSONArray("results");
 
                     for (int i = 0; i < actorsArray.length(); i++) {
@@ -115,9 +249,9 @@ public class SearchByActorActivity extends AppCompatActivity {
                     mProgressBar.setVisibility(View.GONE);
 
 
-                }catch (JSONException e) {
+                } catch (JSONException e) {
                     e.printStackTrace();
-                    Snackbar.make(findViewById(R.layout.activity_search_by_actor),getString(R.string.Nothing_found),Snackbar.LENGTH_LONG).show();
+                    Snackbar.make(findViewById(R.layout.activity_search_by_actor), getString(R.string.Nothing_found), Snackbar.LENGTH_LONG).show();
                 }
 
 
