@@ -3,9 +3,6 @@ package com.aastudio.sarollahi.moviebox.Activities;
 import android.app.ActionBar;
 import android.content.Intent;
 import android.net.Uri;
-import android.support.design.widget.Snackbar;
-import android.support.design.widget.TabLayout;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -13,6 +10,8 @@ import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.aastudio.sarollahi.moviebox.Data.DbHelper;
 import com.aastudio.sarollahi.moviebox.Data.ViewPagerAdapter;
@@ -31,7 +30,10 @@ import com.android.volley.VolleyLog;
 import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.Volley;
 import com.github.ivbaranov.mfb.MaterialFavoriteButton;
+import com.google.android.material.snackbar.Snackbar;
+import com.google.android.material.tabs.TabLayout;
 import com.squareup.picasso.Picasso;
+
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -90,17 +92,9 @@ public class TvDetailActivity extends AppCompatActivity {
         TvShow tvShow = (TvShow) getIntent().getSerializableExtra("movie");
         tvId = tvShow.getMovieId();
 
-        if(dbHelper.getmovie(tvId) == 0){
-            seenbtn = false;
-        }else {
-            seenbtn = true;
-        }
+        seenbtn = dbHelper.getmovie(tvId) != 0;
 
-        if(dbHelper.getfav(tvId) == 0){
-            favbtn = false;
-        }else {
-            favbtn = true;
-        }
+        favbtn = dbHelper.getfav(tvId) != 0;
 
         setUpUI();
         getTvDetails(tvId);
@@ -118,7 +112,7 @@ public class TvDetailActivity extends AppCompatActivity {
         });
 
 
-        MaterialFavoriteButton favorite = (MaterialFavoriteButton) v.findViewById(R.id.favorite_button);
+        MaterialFavoriteButton favorite = v.findViewById(R.id.favorite_button);
         favorite.setFavorite(favbtn, false);
         favorite.setOnFavoriteChangeListener(
                 new MaterialFavoriteButton.OnFavoriteChangeListener() {
@@ -132,7 +126,7 @@ public class TvDetailActivity extends AppCompatActivity {
                     }
                 });
 
-        MaterialFavoriteButton seen = (MaterialFavoriteButton) v.findViewById(R.id.seen_button);
+        MaterialFavoriteButton seen = v.findViewById(R.id.seen_button);
         seen.setFavorite(seenbtn, false);
         seen.setOnFavoriteChangeListener(
                 new MaterialFavoriteButton.OnFavoriteChangeListener() {
@@ -146,16 +140,16 @@ public class TvDetailActivity extends AppCompatActivity {
                     }
                 });
 
-        play_button = (ImageButton) v.findViewById(R.id.play_button);
+        play_button = v.findViewById(R.id.play_button);
         play_button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                if (!trailerText.getText().toString().equals("")){
-                    Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse("https://www.youtube.com/watch?v="+ trailerText.getText().toString()));
+                if (!trailerText.getText().toString().equals("")) {
+                    Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse("https://www.youtube.com/watch?v=" + trailerText.getText().toString()));
                     intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
                     intent.putExtra("VIDEO_ID", trailerText.getText().toString());
                     startActivity(intent);
-                }else {
+                } else {
                     play_button.setBackground(getApplicationContext().getResources().getDrawable(R.drawable.baseline_tv_off_24));
                 }
 
@@ -256,7 +250,7 @@ public class TvDetailActivity extends AppCompatActivity {
                                 movieimage = similar.getString("file_path");
                                 View view = mInflater.inflate(R.layout.index_movie_images_gallery,
                                         iGallery, false);
-                                ImageView movieimg = (ImageView) view
+                                ImageView movieimg = view
                                         .findViewById(R.id.id_index_movie_images);
                                 Picasso.get()
                                         .load("https://image.tmdb.org/t/p/w300" + movieimage)
@@ -290,10 +284,9 @@ public class TvDetailActivity extends AppCompatActivity {
                             JSONObject play = video.getJSONObject(0);
                             trailerText.setText(play.getString("key"));
 
-                        }else {
+                        } else {
                             trailerText.setText("");
                         }
-
 
 
                     } catch (JSONException e) {
@@ -366,6 +359,6 @@ public class TvDetailActivity extends AppCompatActivity {
         runTime = findViewById(R.id.movieTimeIDDets);
         genre = findViewById(R.id.movieGenreIDDet);
         poster = findViewById(R.id.moviePoster);
-        iGallery = (LinearLayout) findViewById(R.id.movie_images_gallery);
+        iGallery = findViewById(R.id.movie_images_gallery);
     }
 }
