@@ -2,8 +2,6 @@ package com.aastudio.sarollahi.moviebox.Data;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,15 +10,17 @@ import android.view.animation.AnimationUtils;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.aastudio.sarollahi.moviebox.Activities.BigImageActivity;
-import com.aastudio.sarollahi.moviebox.Activities.SearchMovieByActorIdActivity;
 import com.aastudio.sarollahi.moviebox.Model.Actor;
 import com.android.sarollahi.moviebox.R;
 import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
-public class PersonImagesRecyclerViewAdapter extends RecyclerView.Adapter<PersonImagesRecyclerViewAdapter.ViewHolder>{
+public class PersonImagesRecyclerViewAdapter extends RecyclerView.Adapter<PersonImagesRecyclerViewAdapter.ViewHolder> {
 
     private Context context;
     private List<Actor> actorList;
@@ -37,7 +37,6 @@ public class PersonImagesRecyclerViewAdapter extends RecyclerView.Adapter<Person
                 .inflate(R.layout.person_tile, parent, false);
 
 
-
         return new ViewHolder(view, context);
     }
 
@@ -45,7 +44,7 @@ public class PersonImagesRecyclerViewAdapter extends RecyclerView.Adapter<Person
     public void onBindViewHolder(PersonImagesRecyclerViewAdapter.ViewHolder holder, int position) {
 
         final Actor actor = actorList.get(position);
-        String posterLink =  "https://image.tmdb.org/t/p/w185"+actor.getPoster();
+        String posterLink = "https://image.tmdb.org/t/p/w185" + actor.getPoster();
 
         holder.title.setText(actor.getName());
 
@@ -62,17 +61,13 @@ public class PersonImagesRecyclerViewAdapter extends RecyclerView.Adapter<Person
             public void onClick(View v) {
 
                 Intent intent = new Intent(context, BigImageActivity.class);
-                intent.putExtra("actor",actor.getPoster());
+                intent.putExtra("actor", actor.getPoster());
                 intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
 
                 context.startActivity(intent);
 
             }
         });
-
-
-
-
 
 
     }
@@ -82,20 +77,20 @@ public class PersonImagesRecyclerViewAdapter extends RecyclerView.Adapter<Person
         return actorList.size();
     }
 
-    public class ViewHolder extends  RecyclerView.ViewHolder implements View.OnClickListener{
+    public class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
 
         TextView title;
         ImageView poster;
-        CardView cardView ;
+        CardView cardView;
 
 
         public ViewHolder(View itemView, final Context ctx) {
             super(itemView);
             context = ctx;
 
-            title = (TextView) itemView.findViewById(R.id.movieTitleID);
-            poster = (ImageView) itemView.findViewById(R.id.movieImageID);
-            cardView = (CardView) itemView.findViewById(R.id.cardview);
+            title = itemView.findViewById(R.id.movieTitleID);
+            poster = itemView.findViewById(R.id.movieImageID);
+            cardView = itemView.findViewById(R.id.cardview);
 
         }
 
