@@ -2,8 +2,6 @@ package com.aastudio.sarollahi.moviebox.Data;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -11,14 +9,18 @@ import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.widget.ImageView;
 import android.widget.TextView;
-import com.aastudio.sarollahi.moviebox.Activities.SearchMovieByActorIdActivity;
+
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.aastudio.sarollahi.moviebox.Activities.PersonActivity;
 import com.aastudio.sarollahi.moviebox.Model.Actor;
 import com.android.sarollahi.moviebox.R;
 import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
-public class PersonRecyclerViewAdapter extends RecyclerView.Adapter<PersonRecyclerViewAdapter.ViewHolder>{
+public class PersonRecyclerViewAdapter extends RecyclerView.Adapter<PersonRecyclerViewAdapter.ViewHolder> {
 
     private Context context;
     private List<Actor> actorList;
@@ -35,7 +37,6 @@ public class PersonRecyclerViewAdapter extends RecyclerView.Adapter<PersonRecycl
                 .inflate(R.layout.person_tile, parent, false);
 
 
-
         return new ViewHolder(view, context);
     }
 
@@ -43,7 +44,7 @@ public class PersonRecyclerViewAdapter extends RecyclerView.Adapter<PersonRecycl
     public void onBindViewHolder(PersonRecyclerViewAdapter.ViewHolder holder, int position) {
 
         final Actor actor = actorList.get(position);
-        String posterLink =  "https://image.tmdb.org/t/p/w185"+actor.getPoster();
+        String posterLink = "https://image.tmdb.org/t/p/w185" + actor.getPoster();
 
         holder.title.setText(actor.getName());
 
@@ -59,18 +60,17 @@ public class PersonRecyclerViewAdapter extends RecyclerView.Adapter<PersonRecycl
             @Override
             public void onClick(View v) {
 
-                Intent intent = new Intent(context, SearchMovieByActorIdActivity.class);
-                intent.putExtra("actor",actor);
+                // Intent intent = new Intent(context, SearchMovieByActorIdActivity.class);
+                Intent intent = new Intent(context, PersonActivity.class);
+                intent.putExtra("personId", actor.getActorId());
+                intent.putExtra("personName", actor.getName());
+                //intent.putExtra("actor", actor);
                 intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
 
                 context.startActivity(intent);
 
             }
         });
-
-
-
-
 
 
     }
@@ -80,20 +80,20 @@ public class PersonRecyclerViewAdapter extends RecyclerView.Adapter<PersonRecycl
         return actorList.size();
     }
 
-    public class ViewHolder extends  RecyclerView.ViewHolder implements View.OnClickListener{
+    public class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
 
         TextView title;
         ImageView poster;
-        CardView cardView ;
+        CardView cardView;
 
 
         public ViewHolder(View itemView, final Context ctx) {
             super(itemView);
             context = ctx;
 
-            title = (TextView) itemView.findViewById(R.id.movieTitleID);
-            poster = (ImageView) itemView.findViewById(R.id.movieImageID);
-            cardView = (CardView) itemView.findViewById(R.id.cardview);
+            title = itemView.findViewById(R.id.movieTitleID);
+            poster = itemView.findViewById(R.id.movieImageID);
+            cardView = itemView.findViewById(R.id.cardview);
 
         }
 
