@@ -1,8 +1,6 @@
 package com.aastudio.sarollahi.moviebox.Data;
 
 import android.content.Context;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -10,13 +8,16 @@ import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.widget.TextView;
 
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.aastudio.sarollahi.moviebox.Model.Review;
 import com.android.sarollahi.moviebox.R;
 import com.uncopt.android.widget.text.justify.JustifiedTextView;
 
 import java.util.List;
 
-public class ReviewRecyclerViewAdapter extends RecyclerView.Adapter<ReviewRecyclerViewAdapter.ViewHolder>{
+public class ReviewRecyclerViewAdapter extends RecyclerView.Adapter<ReviewRecyclerViewAdapter.ViewHolder> {
 
     private Context context;
     private List<Review> reviewList;
@@ -31,7 +32,6 @@ public class ReviewRecyclerViewAdapter extends RecyclerView.Adapter<ReviewRecycl
 
         View view = LayoutInflater.from(parent.getContext())
                 .inflate(R.layout.review_row, parent, false);
-
 
 
         return new ViewHolder(view, context);
@@ -55,20 +55,20 @@ public class ReviewRecyclerViewAdapter extends RecyclerView.Adapter<ReviewRecycl
         return reviewList.size();
     }
 
-    public class ViewHolder extends  RecyclerView.ViewHolder implements View.OnClickListener{
+    public class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
 
         TextView name;
         TextView text;
-        CardView cardView ;
+        CardView cardView;
 
 
         public ViewHolder(View itemView, final Context ctx) {
             super(itemView);
             context = ctx;
 
-            name = (TextView) itemView.findViewById(R.id.id_index_reviews_name);
+            name = itemView.findViewById(R.id.id_index_reviews_name);
             text = (JustifiedTextView) itemView.findViewById(R.id.id_index_reviews_text);
-            cardView = (CardView) itemView.findViewById(R.id.cardview);
+            cardView = itemView.findViewById(R.id.cardview);
 
         }
 
