@@ -4,14 +4,15 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.aastudio.sarollahi.moviebox.Activities.DetailActivity;
 import com.aastudio.sarollahi.moviebox.Activities.SeenMoviesActivity;
@@ -21,7 +22,7 @@ import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
-public class SeenMoviesRecyclerViewAdapter extends RecyclerView.Adapter<SeenMoviesRecyclerViewAdapter.ViewHolder>{
+public class SeenMoviesRecyclerViewAdapter extends RecyclerView.Adapter<SeenMoviesRecyclerViewAdapter.ViewHolder> {
 
     private Context context;
     private List<Movie> movieList;
@@ -42,7 +43,6 @@ public class SeenMoviesRecyclerViewAdapter extends RecyclerView.Adapter<SeenMovi
                 .inflate(R.layout.mylist_row, parent, false);
 
 
-
         return new ViewHolder(view, context);
     }
 
@@ -50,14 +50,14 @@ public class SeenMoviesRecyclerViewAdapter extends RecyclerView.Adapter<SeenMovi
     public void onBindViewHolder(SeenMoviesRecyclerViewAdapter.ViewHolder holder, int position) {
 
         final Movie movie = movieList.get(position);
-        String posterLink =  movie.getPoster();
+        String posterLink = movie.getPoster();
 
         holder.title.setText(movie.getTitle());
         String release_date = movie.getYear();
-        if (release_date.equals("")){
+        if (release_date.equals("")) {
             myear = "--";
-        }else {
-            myear = release_date.substring(0,4);
+        } else {
+            myear = release_date.substring(0, 4);
         }
 
         holder.year.setText(myear + " | " + movie.getOriginalLanguage().toUpperCase());
@@ -75,7 +75,7 @@ public class SeenMoviesRecyclerViewAdapter extends RecyclerView.Adapter<SeenMovi
             public void onClick(View v) {
 
                 Intent intent = new Intent(context, DetailActivity.class);
-                intent.putExtra("movie",movie);
+                intent.putExtra("movie", movie);
                 intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
 
                 context.startActivity(intent);
@@ -98,7 +98,7 @@ public class SeenMoviesRecyclerViewAdapter extends RecyclerView.Adapter<SeenMovi
                         intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
 
                         context.startActivity(intent);
-                        ((Activity)context).finish();
+                        ((Activity) context).finish();
                     }
                 });
                 dialog.setButton(DialogInterface.BUTTON_NEGATIVE, "No", new DialogInterface.OnClickListener() {
@@ -119,26 +119,26 @@ public class SeenMoviesRecyclerViewAdapter extends RecyclerView.Adapter<SeenMovi
         return movieList.size();
     }
 
-    public class ViewHolder extends  RecyclerView.ViewHolder implements View.OnClickListener{
+    public class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
 
         TextView title;
         TextView year;
         TextView overview;
         TextView addedOn;
         ImageView poster;
-        CardView cardView ;
+        CardView cardView;
 
 
         public ViewHolder(View itemView, final Context ctx) {
             super(itemView);
             context = ctx;
 
-            title = (TextView) itemView.findViewById(R.id.movieTitleID);
-            year = (TextView) itemView.findViewById(R.id.movieReleaseID);
-            overview = (TextView) itemView.findViewById(R.id.movieOverview);
-            addedOn = (TextView) itemView.findViewById(R.id.movieaddedID);
-            poster = (ImageView) itemView.findViewById(R.id.movieImageID);
-            cardView = (CardView) itemView.findViewById(R.id.cardview);
+            title = itemView.findViewById(R.id.movieTitleID);
+            year = itemView.findViewById(R.id.movieReleaseID);
+            overview = itemView.findViewById(R.id.movieOverview);
+            addedOn = itemView.findViewById(R.id.movieaddedID);
+            poster = itemView.findViewById(R.id.movieImageID);
+            cardView = itemView.findViewById(R.id.cardview);
 
         }
 
