@@ -4,14 +4,15 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.aastudio.sarollahi.moviebox.Activities.SeenTvActivity;
 import com.aastudio.sarollahi.moviebox.Activities.TvDetailActivity;
@@ -21,7 +22,7 @@ import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
-public class SeenTvRecyclerViewAdapter extends RecyclerView.Adapter<SeenTvRecyclerViewAdapter.ViewHolder>{
+public class SeenTvRecyclerViewAdapter extends RecyclerView.Adapter<SeenTvRecyclerViewAdapter.ViewHolder> {
 
     private Context context;
     private List<TvShow> tvShowList;
@@ -39,7 +40,6 @@ public class SeenTvRecyclerViewAdapter extends RecyclerView.Adapter<SeenTvRecycl
                 .inflate(R.layout.mylist_row, parent, false);
 
 
-
         return new ViewHolder(view, context);
     }
 
@@ -47,14 +47,14 @@ public class SeenTvRecyclerViewAdapter extends RecyclerView.Adapter<SeenTvRecycl
     public void onBindViewHolder(SeenTvRecyclerViewAdapter.ViewHolder holder, int position) {
 
         final TvShow tvShow = tvShowList.get(position);
-        String posterLink =  tvShow.getPoster();
+        String posterLink = tvShow.getPoster();
 
         holder.title.setText(tvShow.getTitle());
         String release_date = tvShow.getYear();
-        if (release_date.equals("")){
+        if (release_date.equals("")) {
             myear = "--";
-        }else {
-            myear = release_date.substring(0,4);
+        } else {
+            myear = release_date.substring(0, 4);
         }
 
         holder.year.setText(myear + " | " + tvShow.getOriginalLanguage().toUpperCase());
@@ -72,7 +72,7 @@ public class SeenTvRecyclerViewAdapter extends RecyclerView.Adapter<SeenTvRecycl
             public void onClick(View v) {
 
                 Intent intent = new Intent(context, TvDetailActivity.class);
-                intent.putExtra("movie",tvShow);
+                intent.putExtra("movie", tvShow);
                 intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
 
                 context.startActivity(intent);
@@ -95,7 +95,7 @@ public class SeenTvRecyclerViewAdapter extends RecyclerView.Adapter<SeenTvRecycl
                         intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
 
                         context.startActivity(intent);
-                        ((Activity)context).finish();
+                        ((Activity) context).finish();
                     }
                 });
                 dialog.setButton(DialogInterface.BUTTON_NEGATIVE, "No", new DialogInterface.OnClickListener() {
@@ -116,26 +116,26 @@ public class SeenTvRecyclerViewAdapter extends RecyclerView.Adapter<SeenTvRecycl
         return tvShowList.size();
     }
 
-    public class ViewHolder extends  RecyclerView.ViewHolder implements View.OnClickListener{
+    public class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
 
         TextView title;
         TextView year;
         TextView overview;
         TextView addedOn;
         ImageView poster;
-        CardView cardView ;
+        CardView cardView;
 
 
         public ViewHolder(View itemView, final Context ctx) {
             super(itemView);
             context = ctx;
 
-            title = (TextView) itemView.findViewById(R.id.movieTitleID);
-            year = (TextView) itemView.findViewById(R.id.movieReleaseID);
-            overview = (TextView) itemView.findViewById(R.id.movieOverview);
-            addedOn = (TextView) itemView.findViewById(R.id.movieaddedID);
-            poster = (ImageView) itemView.findViewById(R.id.movieImageID);
-            cardView = (CardView) itemView.findViewById(R.id.cardview);
+            title = itemView.findViewById(R.id.movieTitleID);
+            year = itemView.findViewById(R.id.movieReleaseID);
+            overview = itemView.findViewById(R.id.movieOverview);
+            addedOn = itemView.findViewById(R.id.movieaddedID);
+            poster = itemView.findViewById(R.id.movieImageID);
+            cardView = itemView.findViewById(R.id.cardview);
 
         }
 
