@@ -2,13 +2,14 @@ package com.aastudio.sarollahi.moviebox.Data;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.aastudio.sarollahi.moviebox.Activities.TvDetailActivity;
 import com.aastudio.sarollahi.moviebox.Model.TvShow;
@@ -17,7 +18,7 @@ import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
-public class TvRecyclerViewAdapter extends RecyclerView.Adapter<TvRecyclerViewAdapter.ViewHolder>{
+public class TvRecyclerViewAdapter extends RecyclerView.Adapter<TvRecyclerViewAdapter.ViewHolder> {
 
     private Context context;
     private List<TvShow> tvShowList;
@@ -35,7 +36,6 @@ public class TvRecyclerViewAdapter extends RecyclerView.Adapter<TvRecyclerViewAd
                 .inflate(R.layout.movie_row, parent, false);
 
 
-
         return new ViewHolder(view, context);
     }
 
@@ -43,16 +43,16 @@ public class TvRecyclerViewAdapter extends RecyclerView.Adapter<TvRecyclerViewAd
     public void onBindViewHolder(TvRecyclerViewAdapter.ViewHolder holder, int position) {
 
         final TvShow movie = tvShowList.get(position);
-        String posterLink =  movie.getPoster();
+        String posterLink = movie.getPoster();
 
         holder.title.setText(movie.getTitle());
         String release_date = movie.getYear();
-        if (release_date.equals("")){
+        if (release_date.equals("")) {
             myear = "--";
-        }else {
-            myear = release_date.substring(0,4);
+        } else {
+            myear = release_date.substring(0, 4);
         }
-        holder.year.setText(myear+ " | " + movie.getOriginalLanguage().toUpperCase());
+        holder.year.setText(myear + " | " + movie.getOriginalLanguage().toUpperCase());
         holder.overview.setText(movie.getPlot());
 
         Picasso.get()
@@ -66,16 +66,12 @@ public class TvRecyclerViewAdapter extends RecyclerView.Adapter<TvRecyclerViewAd
             public void onClick(View v) {
 
                 Intent intent = new Intent(context, TvDetailActivity.class);
-                intent.putExtra("movie",movie);
+                intent.putExtra("movie", movie);
 
                 context.startActivity(intent);
 
             }
         });
-
-
-
-
 
 
     }
@@ -85,24 +81,24 @@ public class TvRecyclerViewAdapter extends RecyclerView.Adapter<TvRecyclerViewAd
         return tvShowList.size();
     }
 
-    public class ViewHolder extends  RecyclerView.ViewHolder implements View.OnClickListener{
+    public class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
 
         TextView title;
         TextView year;
         TextView overview;
         ImageView poster;
-        CardView cardView ;
+        CardView cardView;
 
 
         public ViewHolder(View itemView, final Context ctx) {
             super(itemView);
             context = ctx;
 
-            title = (TextView) itemView.findViewById(R.id.movieTitleID);
-            year = (TextView) itemView.findViewById(R.id.movieReleaseID);
-            overview = (TextView) itemView.findViewById(R.id.movieOverview);
-            poster = (ImageView) itemView.findViewById(R.id.movieImageID);
-            cardView = (CardView) itemView.findViewById(R.id.cardview);
+            title = itemView.findViewById(R.id.movieTitleID);
+            year = itemView.findViewById(R.id.movieReleaseID);
+            overview = itemView.findViewById(R.id.movieOverview);
+            poster = itemView.findViewById(R.id.movieImageID);
+            cardView = itemView.findViewById(R.id.cardview);
 
         }
 
