@@ -1,15 +1,16 @@
 package com.aastudio.sarollahi.moviebox.Fragments;
 
+import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.aastudio.sarollahi.moviebox.Data.MovieRecyclerViewAdapter;
 import com.aastudio.sarollahi.moviebox.Model.Actor;
@@ -22,23 +23,26 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.Volley;
+import com.google.android.material.snackbar.Snackbar;
+
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
+
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 public class FragmentActorMovies extends Fragment {
 
     View v;
-
+    LinearLayoutManager manager;
     private RecyclerView recyclerView;
     private MovieRecyclerViewAdapter movieRecyclerViewAdapter;
     private List<Movie> movieList;
     private RequestQueue queue;
     private Actor actor;
     private String actorId;
-    LinearLayoutManager manager;
 
     public FragmentActorMovies() {
     }
@@ -46,9 +50,9 @@ public class FragmentActorMovies extends Fragment {
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        v = inflater.inflate(R.layout.movies_fragment,container,false);
+        v = inflater.inflate(R.layout.movies_fragment, container, false);
 
-        recyclerView = (RecyclerView) v.findViewById(R.id.recyclerViewMovies);
+        recyclerView = v.findViewById(R.id.recyclerViewMovies);
         recyclerView.setHasFixedSize(true);
         movieRecyclerViewAdapter = new MovieRecyclerViewAdapter(getContext(), movieList);
         manager = new LinearLayoutManager(getContext());
@@ -62,69 +66,72 @@ public class FragmentActorMovies extends Fragment {
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        queue = Volley.newRequestQueue(getActivity());
+        queue = Volley.newRequestQueue(Objects.requireNonNull(getActivity()));
 
         actor = (Actor) getActivity().getIntent().getSerializableExtra("actor");
+        assert actor != null;
         actorId = actor.getActorId();
 
         movieList = new ArrayList<>();
 
-        // getMovies(search);
-        movieList = getMovies(actorId);
+        getMovies getMovies = new getMovies();
+        getMovies.execute(actorId);
     }
 
-    private List<Movie> getMovies(String id) {
+    class getMovies extends AsyncTask<String, Void, List<Movie>> {
+        @Override
+        protected List<Movie> doInBackground(String... id) {
+            JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET,
+                    Constants.MAIN_FIND_MOVIE_BY_ACTOR_LEFT + id[0] + Constants.MAIN_FIND_MOVIE_BY_ACTOR_RIGHT, null, new Response.Listener<JSONObject>() {
+                @Override
+                public void onResponse(JSONObject response) {
 
-        JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET,
-                Constants.MAIN_FIND_MOVIE_BY_ACTOR_LEFT + id + Constants.MAIN_FIND_MOVIE_BY_ACTOR_RIGHT,null, new Response.Listener<JSONObject>() {
-            @Override
-            public void onResponse(JSONObject response) {
+                    try {
 
-                try{
+                        JSONArray moviesArray = response.getJSONArray("cast");
 
-                    JSONArray moviesArray = response.getJSONArray("cast");
+                        for (int i = 0; i < moviesArray.length(); i++) {
 
-                    for (int i = 0; i < moviesArray.length(); i++) {
+                            JSONObject movieObj = moviesArray.getJSONObject(i);
 
-                        JSONObject movieObj = moviesArray.getJSONObject(i);
+                            Movie movie = new Movie();
+                            movie.setTitle(movieObj.getString("title"));
+                            if (movieObj.has("release_date")) {
+                                movie.setYear(movieObj.getString("release_date"));
+                            } else {
+                                movie.setYear("");
+                            }
 
-                        Movie movie = new Movie();
-                        movie.setTitle(movieObj.getString("title"));
-                        if (movieObj.has("release_date")){
-                            movie.setYear(movieObj.getString("release_date"));
-                        }else {
-                            movie.setYear("");
+
+                            movie.setOriginalLanguage(movieObj.getString("original_language"));
+                            movie.setPlot(movieObj.getString("overview"));
+                            movie.setPoster("http://image.tmdb.org/t/p/w185" + movieObj.getString("poster_path"));
+                            movie.setMovieId(movieObj.getString("id"));
+
+                            movieList.add(movie);
+
+
                         }
 
+                        movieRecyclerViewAdapter.notifyDataSetChanged();//Important!!
 
-                        movie.setOriginalLanguage(movieObj.getString("original_language"));
-                        movie.setPlot(movieObj.getString("overview"));
-                        movie.setPoster("http://image.tmdb.org/t/p/w185"+movieObj.getString("poster_path"));
-                        movie.setMovieId(movieObj.getString("id"));
-
-                        movieList.add(movie);
-
-
+                    } catch (JSONException e) {
+                        e.printStackTrace();
+                        Snackbar.make(getView(), R.string.No_movies_found, Snackbar.LENGTH_LONG).show();
                     }
 
-                    movieRecyclerViewAdapter.notifyDataSetChanged();//Important!!
 
-                }catch (JSONException e) {
-                    e.printStackTrace();
-                    Snackbar.make(getView(),R.string.No_movies_found,Snackbar.LENGTH_LONG).show();
                 }
+            }, new Response.ErrorListener() {
+                @Override
+                public void onErrorResponse(VolleyError error) {
 
+                }
+            });
 
-            }
-        }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
+            queue.add(jsonObjectRequest);
 
-            }
-        });
-
-        queue.add(jsonObjectRequest);
-
-        return movieList;
+            return movieList;
+        }
     }
 }
