@@ -1,20 +1,22 @@
 package com.aastudio.sarollahi.moviebox.Fragments;
 
+import android.app.ProgressDialog;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
+import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.ImageView;
-import android.widget.ProgressBar;
+import android.widget.ScrollView;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import com.aastudio.sarollahi.moviebox.Activities.SearchMovieByActorIdActivity;
 import com.aastudio.sarollahi.moviebox.Model.Actor;
@@ -48,7 +50,8 @@ public class FragmentBiography extends Fragment {
     private ImageView poster;
     private Button searchbyact;
     private Context mContext;
-    private ProgressBar mProgressBar;
+    private String personId;
+    private ScrollView page;
 
     public FragmentBiography() {
     }
@@ -56,7 +59,7 @@ public class FragmentBiography extends Fragment {
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        v = inflater.inflate(R.layout.person_biography_fragment,container,false);
+        v = inflater.inflate(R.layout.person_biography_fragment, container, false);
 
         name = v.findViewById(R.id.actorname);
         birth = v.findViewById(R.id.actorbirth);
@@ -68,7 +71,7 @@ public class FragmentBiography extends Fragment {
         facebook = v.findViewById(R.id.facebookactor);
         instagram = v.findViewById(R.id.instagramactor);
         searchbyact = v.findViewById(R.id.searchMTbyactor);
-        mProgressBar = (ProgressBar) v.findViewById(R.id.progressBar);
+        page = v.findViewById(R.id.bioPage);
         return v;
     }
 
@@ -87,147 +90,156 @@ public class FragmentBiography extends Fragment {
 
         Bundle person = getActivity().getIntent().getExtras();
 
-        String personId = person.getString("personId");
+        personId = person.getString("personId");
 
-
-        getPersonDetails(personId);
+        getPersonDetails getPersonDetails = new getPersonDetails();
+        getPersonDetails.execute(personId);
     }
 
-    private void getPersonDetails(final String personId) {
 
-        getActivity().getWindow().setFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE,
-                WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE);
+    class getPersonDetails extends AsyncTask<String, Void, Void> {
+        ProgressDialog progressDialog;
 
-        JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET
-                , Constants.PERSON_INFO_URL_LEFT + personId + Constants.PERSON_INFO_URL_RIGHT, null, new Response.Listener<JSONObject>() {
-            @Override
-            public void onResponse(final JSONObject response) {
+        @Override
+        protected Void doInBackground(String... id) {
 
-                try {
+            JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET
+                    , Constants.PERSON_INFO_URL_LEFT + id[0] + Constants.PERSON_INFO_URL_RIGHT, null, new Response.Listener<JSONObject>() {
+                @Override
+                public void onResponse(final JSONObject response) {
 
-                    String actname ="";
+                    try {
 
-                    name.setText(response.getString("name"));
-                    if (!response.getString("birthday").equals("null")){
-                        birth.setText(response.getString("birthday"));
-                    }else {
-                        birth.setText("");
-                    }
-                    if (!response.getString("place_of_birth").equals("null")){
-                        birthplace.setText(response.getString("place_of_birth"));
-                    }else {
-                        birthplace.setText("");
-                    }
+                        String actname = "";
 
-                    bio.setText(response.getString("biography"));
-                    Picasso.get().load("https://image.tmdb.org/t/p/h632"+response.getString("profile_path")).placeholder(R.drawable.noimage).into(poster);
-                    if (response.has("name")){
-                        actname = response.getString("name");
-                    }else {
-                        actname = "";
-                    }
-                    final String finalActname = actname;
-                    searchbyact.setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
-                            Actor act = new Actor();
-                            act.setActorId(personId);
-                            act.setName(finalActname);
-                            Intent intent = new Intent(getActivity(),SearchMovieByActorIdActivity.class);
-                            intent.putExtra("actor",act);
-                            startActivity(intent);
+                        name.setText(response.getString("name"));
+                        if (!response.getString("birthday").equals("null")) {
+                            birth.setText(response.getString("birthday"));
+                        } else {
+                            birth.setText("");
                         }
-                    });
-                    if (!response.isNull("homepage")){
+                        if (!response.getString("place_of_birth").equals("null")) {
+                            birthplace.setText(response.getString("place_of_birth"));
+                        } else {
+                            birthplace.setText("");
+                        }
 
-                        final String finalWeb = response.getString("homepage");
-                        website.setVisibility(View.VISIBLE);
-                        website.setOnClickListener(new View.OnClickListener() {
+                        bio.setText(response.getString("biography"));
+                        Picasso.get().load("https://image.tmdb.org/t/p/h632" + response.getString("profile_path")).placeholder(R.drawable.noimage).into(poster);
+                        if (response.has("name")) {
+                            actname = response.getString("name");
+                        } else {
+                            actname = "";
+                        }
+                        final String finalActname = actname;
+                        searchbyact.setOnClickListener(new View.OnClickListener() {
                             @Override
                             public void onClick(View v) {
-                                Intent intent = new Intent();
-                                intent.setAction(Intent.ACTION_VIEW);
-                                intent.addCategory(Intent.CATEGORY_BROWSABLE);
-                                intent.setData(Uri.parse(finalWeb));
+                                Actor act = new Actor();
+                                act.setActorId(personId);
+                                act.setName(finalActname);
+                                Intent intent = new Intent(getActivity(), SearchMovieByActorIdActivity.class);
+                                intent.putExtra("actor", act);
                                 startActivity(intent);
                             }
                         });
+                        if (!response.isNull("homepage")) {
 
-                    }
-                    try {
-                        JSONObject exlink = response.getJSONObject("external_ids");
-
-
-
-                        if (!exlink.isNull("imdb_id")){
-                            imdb.setVisibility(View.VISIBLE);
-                            final String im = exlink.getString("imdb_id");
-                            imdb.setOnClickListener(new View.OnClickListener() {
+                            final String finalWeb = response.getString("homepage");
+                            website.setVisibility(View.VISIBLE);
+                            website.setOnClickListener(new View.OnClickListener() {
                                 @Override
                                 public void onClick(View v) {
                                     Intent intent = new Intent();
                                     intent.setAction(Intent.ACTION_VIEW);
                                     intent.addCategory(Intent.CATEGORY_BROWSABLE);
-                                    intent.setData(Uri.parse("https://www.imdb.com/name/"+ im+"/"));
+                                    intent.setData(Uri.parse(finalWeb));
                                     startActivity(intent);
                                 }
                             });
 
                         }
-                        if (!exlink.isNull("facebook_id")){
-
-                            final String face = exlink.getString("facebook_id");
-                            facebook.setOnClickListener(new View.OnClickListener() {
-                                @Override
-                                public void onClick(View v) {
-                                    Intent intent = new Intent();
-                                    intent.setAction(Intent.ACTION_VIEW);
-                                    intent.addCategory(Intent.CATEGORY_BROWSABLE);
-                                    intent.setData(Uri.parse("https://www.facebook.com/"+ face+"/"));
-                                    startActivity(intent);
-                                }
-                            });
-
-                        }
-                        if (!exlink.isNull("instagram_id")){
-
-                            final String insta = exlink.getString("instagram_id");
-                            instagram.setOnClickListener(new View.OnClickListener() {
-                                @Override
-                                public void onClick(View v) {
-                                    Intent intent = new Intent();
-                                    intent.setAction(Intent.ACTION_VIEW);
-                                    intent.addCategory(Intent.CATEGORY_BROWSABLE);
-                                    intent.setData(Uri.parse("https://www.instagram.com/"+ insta+"/"));
-                                    startActivity(intent);
-                                }
-                            });
-                        }
-
                         try {
-                            getActivity().getWindow().clearFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE);
-                        } catch (Exception e){
+                            JSONObject exlink = response.getJSONObject("external_ids");
 
+
+                            if (!exlink.isNull("imdb_id")) {
+                                imdb.setVisibility(View.VISIBLE);
+                                final String im = exlink.getString("imdb_id");
+                                imdb.setOnClickListener(new View.OnClickListener() {
+                                    @Override
+                                    public void onClick(View v) {
+                                        Intent intent = new Intent();
+                                        intent.setAction(Intent.ACTION_VIEW);
+                                        intent.addCategory(Intent.CATEGORY_BROWSABLE);
+                                        intent.setData(Uri.parse("https://www.imdb.com/name/" + im + "/"));
+                                        startActivity(intent);
+                                    }
+                                });
+
+                            }
+                            if (!exlink.isNull("facebook_id")) {
+
+                                final String face = exlink.getString("facebook_id");
+                                facebook.setOnClickListener(new View.OnClickListener() {
+                                    @Override
+                                    public void onClick(View v) {
+                                        Intent intent = new Intent();
+                                        intent.setAction(Intent.ACTION_VIEW);
+                                        intent.addCategory(Intent.CATEGORY_BROWSABLE);
+                                        intent.setData(Uri.parse("https://www.facebook.com/" + face + "/"));
+                                        startActivity(intent);
+                                    }
+                                });
+
+                            }
+                            if (!exlink.isNull("instagram_id")) {
+
+                                final String insta = exlink.getString("instagram_id");
+                                instagram.setOnClickListener(new View.OnClickListener() {
+                                    @Override
+                                    public void onClick(View v) {
+                                        Intent intent = new Intent();
+                                        intent.setAction(Intent.ACTION_VIEW);
+                                        intent.addCategory(Intent.CATEGORY_BROWSABLE);
+                                        intent.setData(Uri.parse("https://www.instagram.com/" + insta + "/"));
+                                        startActivity(intent);
+                                    }
+                                });
+                            }
+
+                        } catch (JSONException e) {
+                            e.printStackTrace();
                         }
-                        mProgressBar.setVisibility(View.GONE);
+
 
                     } catch (JSONException e) {
                         e.printStackTrace();
                     }
 
-
-                } catch (JSONException e) {
-                    e.printStackTrace();
                 }
+            }, new Response.ErrorListener() {
+                @Override
+                public void onErrorResponse(VolleyError error) {
 
-            }
-        }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
+                }
+            });
 
-            }
-        });
+            queue.add(jsonObjectRequest);
+            return null;
+        }
 
-        queue.add(jsonObjectRequest);
+        @Override
+        protected void onPreExecute() {
+            progressDialog = ProgressDialog.show(getContext(), null, "Loading...");
+        }
+
+        @Override
+        protected void onPostExecute(Void result) {
+            progressDialog.dismiss();
+            page.setVisibility(View.VISIBLE);
+        }
     }
+
+
 }
