@@ -1,31 +1,32 @@
 package com.aastudio.sarollahi.moviebox.Fragments;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.aastudio.sarollahi.moviebox.Data.DbHelper;
 import com.aastudio.sarollahi.moviebox.Data.SeenTvRecyclerViewAdapter;
 import com.aastudio.sarollahi.moviebox.Model.TvShow;
 import com.android.sarollahi.moviebox.R;
+
 import java.util.ArrayList;
 import java.util.List;
 
 public class FragmentFavoriteTvShows extends Fragment {
 
     View v;
-
+    LinearLayoutManager manager;
     private RecyclerView recyclerView;
     private SeenTvRecyclerViewAdapter movieRecyclerViewAdapter;
     private List<TvShow> movieList;
     private DbHelper db;
-    LinearLayoutManager manager;
 
 
     public FragmentFavoriteTvShows() {
@@ -34,12 +35,12 @@ public class FragmentFavoriteTvShows extends Fragment {
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        v = inflater.inflate(R.layout.tvshows_fragment,container,false);
+        v = inflater.inflate(R.layout.tvshows_fragment, container, false);
 
 
-        recyclerView = (RecyclerView) v.findViewById(R.id.recyclerViewTvShows);
+        recyclerView = v.findViewById(R.id.recyclerViewTvShows);
         recyclerView.setHasFixedSize(true);
-        movieRecyclerViewAdapter = new SeenTvRecyclerViewAdapter(getContext(), movieList );
+        movieRecyclerViewAdapter = new SeenTvRecyclerViewAdapter(getContext(), movieList);
         manager = new LinearLayoutManager(getContext());
         recyclerView.setLayoutManager(manager);
         recyclerView.setAdapter(movieRecyclerViewAdapter);
