@@ -1,21 +1,23 @@
 package com.aastudio.sarollahi.moviebox.Fragments;
 
 import android.content.Context;
+import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.aastudio.sarollahi.moviebox.Data.ReviewRecyclerViewAdapter;
 import com.aastudio.sarollahi.moviebox.Model.Movie;
 import com.aastudio.sarollahi.moviebox.Model.Review;
-import com.android.sarollahi.moviebox.R;
 import com.aastudio.sarollahi.moviebox.Util.Constants;
+import com.android.sarollahi.moviebox.R;
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.Response;
@@ -41,20 +43,23 @@ public class FragmentReviews extends Fragment {
     private RequestQueue queue;
     private Context mContext;
 
+
     public FragmentReviews() {
     }
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        v = inflater.inflate(R.layout.review_fragment,container,false);
+        v = inflater.inflate(R.layout.review_fragment, container, false);
 
-        recyclerView = (RecyclerView) v.findViewById(R.id.recyclerViewReview);
+        recyclerView = v.findViewById(R.id.recyclerViewReview);
         recyclerView.setHasFixedSize(true);
-        reviewRecyclerViewAdapter = new ReviewRecyclerViewAdapter(getContext(), reviewList );
+        reviewRecyclerViewAdapter = new ReviewRecyclerViewAdapter(getContext(), reviewList);
         recyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
         recyclerView.setAdapter(reviewRecyclerViewAdapter);
         reviewRecyclerViewAdapter.notifyDataSetChanged();
+
+
         return v;
     }
 
@@ -69,8 +74,6 @@ public class FragmentReviews extends Fragment {
         super.onCreate(savedInstanceState);
 
 
-
-
         queue = Volley.newRequestQueue(mContext);
 
         movie = (Movie) getActivity().getIntent().getSerializableExtra("movie");
@@ -78,61 +81,66 @@ public class FragmentReviews extends Fragment {
 
         reviewList = new ArrayList<>();
 
-        reviewList = getreviews(movieId);
+        getReview getReviews = new getReview();
+        getReviews.execute(movieId);
     }
 
-    private List<Review> getreviews(String movieid) {
-        reviewList.clear();
+    class getReview extends AsyncTask<String, Void, List<Review>> {
+        @Override
+        protected List<Review> doInBackground(String... id) {
+            reviewList.clear();
 
-        JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET,
-                Constants.MAIN_INFO_URL_LEFT + movieid + Constants.MAIN_INFO_URL_RIGHT_REVIEWS, null, new Response.Listener<JSONObject>() {
-            @Override
-            public void onResponse(JSONObject response) {
+            JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET,
+                    Constants.MAIN_INFO_URL_LEFT + id[0] + Constants.MAIN_INFO_URL_RIGHT_REVIEWS, null, new Response.Listener<JSONObject>() {
+                @Override
+                public void onResponse(JSONObject response) {
 
-                try {
-                    JSONObject reviewss = response.getJSONObject("reviews");
-                    JSONArray reviews = reviewss.getJSONArray("results");
+                    try {
+                        JSONObject reviewss = response.getJSONObject("reviews");
+                        JSONArray reviews = reviewss.getJSONArray("results");
 
-                    String name = null;
-                    String text = null;
 
-                    if (reviews.length() > 0) {
+                        String name = null;
+                        String text = null;
 
-                        for (int i = 0; i <= reviews.length(); i++) {
+                        if (reviews.length() > 0) {
 
-                            JSONObject rev = reviews.getJSONObject(i);
+                            for (int i = 0; i <= 0; i++) {
+
+                                JSONObject rev = reviews.getJSONObject(i);
+                                Review review = new Review();
+                                review.setName(rev.getString("author") + ":");
+                                review.setText(rev.getString("content"));
+
+                                reviewList.add(review);
+                            }
+
+                            reviewRecyclerViewAdapter.notifyDataSetChanged();
+
+                        } else {
                             Review review = new Review();
-                            review.setName(rev.getString("author")+":");
-                            review.setText(rev.getString("content"));
+                            review.setName("No reviews found");
+                            review.setText("");
 
                             reviewList.add(review);
+
+                            reviewRecyclerViewAdapter.notifyDataSetChanged();
                         }
 
-                        reviewRecyclerViewAdapter.notifyDataSetChanged();//Important!!
-
-                    }else {
-                        Review review = new Review();
-                        review.setName("No reviews found");
-                        review.setText("");
-
-                        reviewList.add(review);
-
-                        reviewRecyclerViewAdapter.notifyDataSetChanged();//Important!!
+                    } catch (JSONException e) {
+                        e.printStackTrace();
                     }
 
-                } catch (JSONException e) {
-                    e.printStackTrace();
+
                 }
+            }, new Response.ErrorListener() {
+                @Override
+                public void onErrorResponse(VolleyError error) {
 
-
-            }
-        }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
-
-            }
-        });
-        queue.add(jsonObjectRequest);
-        return reviewList;
+                }
+            });
+            queue.add(jsonObjectRequest);
+            return reviewList;
+        }
     }
 }
