@@ -1,14 +1,15 @@
 package com.aastudio.sarollahi.moviebox.Fragments;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.aastudio.sarollahi.moviebox.Data.ReviewRecyclerViewAdapter;
 import com.aastudio.sarollahi.moviebox.Model.Review;
@@ -45,14 +46,15 @@ public class FragmentTvReviews extends Fragment {
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        v = inflater.inflate(R.layout.review_fragment,container,false);
+        v = inflater.inflate(R.layout.review_fragment, container, false);
 
-        recyclerView = (RecyclerView) v.findViewById(R.id.recyclerViewReview);
+        recyclerView = v.findViewById(R.id.recyclerViewReview);
         recyclerView.setHasFixedSize(true);
-        reviewRecyclerViewAdapter = new ReviewRecyclerViewAdapter(getContext(), reviewList );
+        reviewRecyclerViewAdapter = new ReviewRecyclerViewAdapter(getContext(), reviewList);
         recyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
         recyclerView.setAdapter(reviewRecyclerViewAdapter);
         reviewRecyclerViewAdapter.notifyDataSetChanged();
+
         return v;
     }
 
@@ -68,6 +70,7 @@ public class FragmentTvReviews extends Fragment {
         reviewList = new ArrayList<>();
 
         reviewList = getreviews(movieId);
+
     }
 
     private List<Review> getreviews(String movieid) {
@@ -87,11 +90,11 @@ public class FragmentTvReviews extends Fragment {
 
                     if (reviews.length() > 0) {
 
-                        for (int i = 0; i <= reviews.length(); i++) {
+                        for (int i = 0; i <= 0; i++) {
 
                             JSONObject rev = reviews.getJSONObject(i);
                             Review review = new Review();
-                            review.setName(rev.getString("author")+":");
+                            review.setName(rev.getString("author") + ":");
                             review.setText(rev.getString("content"));
 
                             reviewList.add(review);
@@ -99,7 +102,7 @@ public class FragmentTvReviews extends Fragment {
 
                         reviewRecyclerViewAdapter.notifyDataSetChanged();//Important!!
 
-                    }else {
+                    } else {
                         Review review = new Review();
                         review.setName("No reviews found");
                         review.setText("");
@@ -124,4 +127,6 @@ public class FragmentTvReviews extends Fragment {
         queue.add(jsonObjectRequest);
         return reviewList;
     }
+
+
 }
