@@ -1,16 +1,16 @@
 package com.aastudio.sarollahi.moviebox.Fragments;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AbsListView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.aastudio.sarollahi.moviebox.Data.TvRecyclerViewAdapter;
 import com.aastudio.sarollahi.moviebox.Model.TvShow;
@@ -22,6 +22,7 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.Volley;
+import com.google.android.material.snackbar.Snackbar;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -32,21 +33,45 @@ import java.util.List;
 
 public class FragmentTvShows extends Fragment {
 
+    private static final int PAGE_START = 1;
     View v;
-
+    String search;
+    String year;
+    LinearLayoutManager manager;
+    int currentItems, totalItems, scrollOutItems;
+    Boolean isScrolling = false;
     private RecyclerView recyclerView;
     private TvRecyclerViewAdapter movieRecyclerViewAdapter;
     private List<TvShow> movieList;
     private RequestQueue queue;
-    String search;
-    String year;
-    LinearLayoutManager manager;
-
-    private static final int PAGE_START = 1;
     private int TOTAL_PAGES = 1;
     private int currentPage = PAGE_START;
-    int currentItems, totalItems, scrollOutItems;
-    Boolean isScrolling = false;
+    private final RecyclerView.OnScrollListener scrollListener = new RecyclerView.OnScrollListener() {
+        @Override
+        public void onScrollStateChanged(@NonNull RecyclerView recyclerView, int newState) {
+            super.onScrollStateChanged(recyclerView, newState);
+            if (newState == AbsListView.OnScrollListener.SCROLL_STATE_TOUCH_SCROLL) {
+                isScrolling = true;
+            }
+        }
+
+        @Override
+        public void onScrolled(@NonNull RecyclerView recyclerView, int dx, int dy) {
+            super.onScrolled(recyclerView, dx, dy);
+
+            currentItems = manager.getChildCount();
+            totalItems = manager.getItemCount();
+            scrollOutItems = manager.findFirstVisibleItemPosition();
+
+            if (isScrolling && (currentItems + scrollOutItems == totalItems) && currentPage < TOTAL_PAGES) {
+
+                isScrolling = false;
+                currentPage++;
+                movieList = getMovies(search, year, currentPage);
+
+            }
+        }
+    };
 
     public FragmentTvShows() {
     }
@@ -54,11 +79,11 @@ public class FragmentTvShows extends Fragment {
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        v = inflater.inflate(R.layout.tvshows_fragment,container,false);
+        v = inflater.inflate(R.layout.tvshows_fragment, container, false);
 
-        recyclerView = (RecyclerView) v.findViewById(R.id.recyclerViewTvShows);
+        recyclerView = v.findViewById(R.id.recyclerViewTvShows);
         recyclerView.setHasFixedSize(true);
-        movieRecyclerViewAdapter = new TvRecyclerViewAdapter(getContext(), movieList );
+        movieRecyclerViewAdapter = new TvRecyclerViewAdapter(getContext(), movieList);
         manager = new LinearLayoutManager(getContext());
         recyclerView.setLayoutManager(manager);
         recyclerView.setAdapter(movieRecyclerViewAdapter);
@@ -80,47 +105,18 @@ public class FragmentTvShows extends Fragment {
         movieList = new ArrayList<>();
 
         // getMovies(search);
-        movieList = getMovies(search,year,currentPage);
+        movieList = getMovies(search, year, currentPage);
     }
 
-    private final RecyclerView.OnScrollListener scrollListener = new RecyclerView.OnScrollListener() {
-        @Override
-        public void onScrollStateChanged(@NonNull RecyclerView recyclerView, int newState) {
-            super.onScrollStateChanged(recyclerView, newState);
-            if(newState == AbsListView.OnScrollListener.SCROLL_STATE_TOUCH_SCROLL)
-            {
-                isScrolling = true;
-            }
-        }
-
-        @Override
-        public void onScrolled(@NonNull RecyclerView recyclerView, int dx, int dy) {
-            super.onScrolled(recyclerView, dx, dy);
-
-            currentItems = manager.getChildCount();
-            totalItems = manager.getItemCount();
-            scrollOutItems = manager.findFirstVisibleItemPosition();
-
-            if(isScrolling && (currentItems + scrollOutItems == totalItems) && currentPage<TOTAL_PAGES)
-            {
-
-                isScrolling = false;
-                currentPage++;
-                movieList = getMovies(search,year,currentPage);
-
-            }
-        }
-    };
-
-    private List<TvShow> getMovies(String name,String year,int page) {
+    private List<TvShow> getMovies(String name, String year, int page) {
 
 
         JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET,
-                Constants.TV_SHOWS + name + Constants.MAIN_URL_RIGHT_tV + year+Constants.URL_PAGE+page,null, new Response.Listener<JSONObject>() {
+                Constants.TV_SHOWS + name + Constants.MAIN_URL_RIGHT_tV + year + Constants.URL_PAGE + page, null, new Response.Listener<JSONObject>() {
             @Override
             public void onResponse(JSONObject response) {
 
-                try{
+                try {
                     TOTAL_PAGES = Integer.parseInt(response.getString("total_pages"));
 
                     JSONArray moviesArray = response.getJSONArray("results");
@@ -135,7 +131,7 @@ public class FragmentTvShows extends Fragment {
 
                         movie.setOriginalLanguage(movieObj.getString("original_language"));
                         movie.setPlot(movieObj.getString("overview"));
-                        movie.setPoster("http://image.tmdb.org/t/p/w185"+movieObj.getString("poster_path"));
+                        movie.setPoster("http://image.tmdb.org/t/p/w185" + movieObj.getString("poster_path"));
                         movie.setMovieId(movieObj.getString("id"));
 
                         movieList.add(movie);
@@ -145,9 +141,9 @@ public class FragmentTvShows extends Fragment {
 
                     movieRecyclerViewAdapter.notifyDataSetChanged();//Important!!
 
-                }catch (JSONException e) {
+                } catch (JSONException e) {
                     e.printStackTrace();
-                    Snackbar.make(getView(), R.string.No_tv_shows_found,Snackbar.LENGTH_LONG).show();
+                    Snackbar.make(getView(), R.string.No_tv_shows_found, Snackbar.LENGTH_LONG).show();
                 }
 
 
