@@ -1,10 +1,9 @@
 package com.aastudio.sarollahi.moviebox.Model;
 
 import java.io.Serializable;
-import java.util.Comparator;
 
-public class Movie implements Serializable{
-    private static  final long id = 1L;
+public class Movie implements Serializable {
+    private static final long id = 1L;
 
     private int ids;
     private String title;
