@@ -1,11 +1,12 @@
 package com.aastudio.sarollahi.moviebox.Util;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.view.ViewPager;
 import android.util.AttributeSet;
 import android.view.View;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.viewpager.widget.ViewPager;
 
 public class CustomViewPager extends ViewPager {
     public CustomViewPager(@NonNull Context context) {
